introduce backend_registry.py
